+++ CHANGELOG.md
@@ -13,7 +13,7 @@ will not have contiguous patch numbers. Initial major and minor releases will be
 in this file without a patch number. Patch version will be included for bug fixes, but
 may not exactly match a publicly released version.
 
-## [Unreleased]
+## [3.3.23] - 2022-07-06
 
 ### Added
 * Adds `ContractNameOverride` property to `NeoContractReference` items to control the name of the generated contract interface (#35)